Setting sample rate works across the board
Sample rate now works in the DeviceControlActivity and LongTerm Classes.

Z_BLE_Debug_Timer sketch now creates a pseudo signal at a constant rate
regardless of the sample rate of the phone. It allows for the
demonstration of the effect choosing a sample rate has.

TODO: Set up the sample rate to sync with the file writing methods so
that regardless of sample rate, text files are still produced every 12
minutes, every hour, and every 6 hours