New name; also many improvements and bug fixes
Name is now “Biohm” a concatenation of Bio-Impedance, where impedance
has been replaced by its unit of measure.

In the long term activity, hitting the navigation drawer button
actually pulls out the navigation bar instead of taking you back to the
previous activity (Scanning).

In the DeviceControlActivityClass, the HorizontalScrollView that housed
the live data graph and the text representation has been replaced with
a regular (vertical) ScrollView. The overall UI is much more cohesive
as a result.

In the DeviceControlActivityClass, the ExpandableListView that used the
Gatt structure of the client has been hidden and replaced by a simple
start button to start collecting data. Should one want to view the Gatt
Structure, there will be an option in the yet to be created settings
activity where the user may peruse it.

Overall, the app is much more polished at this stage.