Added and implemented 40 µs library
Previously there was no difference between a sample rate of 85 and 90
Hz as the previous timer had a 1 ms resolution cast values from 85 - 90
Hz as 11 milliseconds. With a 40 µs, differences between 89 and 90 Hz
can now be found at 111235 and 11111 µs respectively.

NOTE: at 40 µs of resolution, accuracy is not perfect as the interrupt
counts in multiples of 40. Also, since the values gotten from the
inverse of the frequency have to be cast to a long, decimal significant
values are lost.