Improved BLE connectivity at higher sample rates
This was done mostly by adjusting the value for cycles base.

From the AD5933 data sheet, this variable is needed for high “Q”
(quality factor) circuits as these tend to oscillate for long periods
of time, and therefore don’t dampen quickly.

Q is a is dependent on the values of R, L and C in RLC circuits; mostly
the series resistor connected to the inductors and capacitors. Since we
calibrate with a simple resistor, and human tissue possess little to no
inductance to my knowledge, we can set this value to really low numbers
(5 - 10 cycles).

This has no impact at low sample rates (< 50 Hz), but at values higher,
after each sweep is complete, the delay from ignoring successive
measurements hampers the connection quality and utilizes necessary CPU
cycles. Setting cycles_vase to 5 alleviates this issue.