A model for TJ to explore
This is a quick and easy way to see how you expect impedance to change
as a function of the three constitutive components. Since you are
cutting through a hyperspace (Re,Ri,Cm,F) it can be hard to orient your
brain about what’s going on. Maybe this will help.