Create v0.4 for all functions
daBI_v0.4 still needs a lot of work but an if structure was created. I
would rather it be a case structure inside of a function, and I also
don’t want the gross graphing code at the bottom to be anywhere near
eyeballs. So those must work perfectly in order to be relegated to the
background.