MAJOR LIBRARY UPDATE
Due to space constraints on the ATMEGA 1280, verbosity in library
functions and sparse documentation, the entire AD5933 library has been
overhauled to keep only essential library functions.

These functions are all documented in a Javadoc / Doxygen type format
in the .cpp file. This should make the library easy to use for anyone
perusing the repository or new comers to the project.

As a result, most sketches in the project are now deprecated, but kept
as legacy sketches. The sketches are still valid, but functions that
have been renamed will need to be updated to get them working.

Some of these sketches will be restored or rewritten in a single sketch
that lets the user toggle the functions from the serial monitor.